changed dtype from float32 to float16

+++ src/data_processing.py
@@ -265,11 +265,11 @@ class RAMAug(Dataset):
         type(idx)
 
         original_image = np.array(
-            self.original_dataset[idx: idx + 1, :, :], copy=True, dtype=np.float32
+            self.original_dataset[idx: idx + 1, :, :], copy=True, dtype=np.float16
         )
         #print("or_img: ", original_image.transpose(0,3,1,2).shape)
 
-        label = np.array(self.labels[idx], copy=True, dtype=np.float32)
+        label = np.array(self.labels[idx], copy=True, dtype=np.float16)
 
         return_dict = {"original": original_image, "label": label}
 
@@ -278,7 +278,7 @@ class RAMAug(Dataset):
             aug_image = np.array(
                 self.aug_datasets[aug_version][idx: idx+1, :, :],
                 copy=True,
-                dtype=np.float32,
+                dtype=np.float16,
             )
             #print("aug_img: ", aug_image.transpose(0,3,1,2).squeeze().shape)
             return_dict["aug"] = aug_image
